(1) moved "sort mode", "sort order" settings to song selection screen (2) added filtering songs by attributes (3) added sort by song Lv, remove sort by filename

+++ core/src/com/fteams/sstrain/entities/Metadata.java
@@ -4,7 +4,7 @@ import com.fteams.sstrain.util.SongUtils;
 
 public class Metadata extends BaseMetadata{
     public String difficultyName;
-    public Long difficulty;
+    public Integer difficulty;
 
     public int compareTo(Metadata metadata) {
         if (songName.equals(metadata.songName))
